requirements.txt
streamlit
mobile-sam
opencv-python
scipy
numpy
matplotlib
torch
torchvision
Pillow
